Add SoGo MySQL support & remove Postgresql
This build should finally work.   There was a docker build error each time I added the package sope4.9-gdl1-mysql.   This was due to the fact that the original upstream "from baseimage" command was grabing "latest" and not a specific version tag of the Ubuntu build.  This caused the baseimage to build Ubuntu 16.04 LTS (Xenial) while trying to install the Trusty version (Ubuntu 14.04LTS) of SoGO which then resulted in a fatal dependancy error.